mg100: battery: fix various issues found during the review
this change fixes various issues found during the review.

-whitespace and spelling issues throughout
-reverting ble_power_service changes and instead using a new
ble_battery_service. there is now a set of UUIDs specifically for
the battery data. no external power voltages will be reported because
the hardware has no way of measuring that. only battery voltage will
be reported.
-fixed the low battery alarm behavior to only alarm if the battery
is low and not externally powered.
-updated the nv writes to just writing instead of reading first. there
is already protection enabled to prevent writes of identical data to
flash.
-fixed the order of init so that functions will not be used before
the ble service is initialized
-updated the ble readme to reflect changes to the battery chars

Bug 16620