fix the gateway sensor array parser to fix handling of max sensors
this change updates the parsing algorithm to fix an issue handling
a sensory array from the server with exactly the maxiumum number
of sensors. instead of rejecting the entire request it will simply
parse out the number of sensors the gateway can handle and ignore
the rest. this also brings the mg100 to the same functionality of
the Pinnacle 100 OOB demo.

Bug 17543