Generate firmware build ID
Use cmake to generate a firmware build ID that
can be included in the firmware version.
This will make the firmware aware of its build ID and allow us to target release versions easier.

Also bump the version to 6.3.99 to indicate active development and indicate we want the next release version to be 6.4.x or greater.

bug 21481