LwM2M watchdog: change system reboot behavior
Change the LwM2M watchdog to not reboot the system too often.
Instead, force a LwM2M disconnect when the watchdog
expires to cleanup the connection.

Add a new supervision watchdog with a very long expiration.
If this watchdog expires, the system will reboot.

Add a Memfault trace to track when the connection cleanup
watchdog expires.

bug 22199